completing the tests

+++ dataAccess/repositories/commentRoutesRepo.mjs
@@ -1,18 +1,22 @@
-import { postModel } from "../models/postModel.mjs";
-import { userModel } from "../models/userModel.mjs";
-import { commentModel } from "../models/commentModel.mjs";
+import { postModel } from "../models/postModel";
+import { userModel } from "../models/userModel";
+import { commentModel } from "../models/commentModel";
 
 export default {
-    
   addComment: async (req, res) => {
     const { userId, postId, body } = req.body;
+    const validation = userId && postId && body;
     try {
-      const newComment = await commentModel.create({
-        userId: userId,
-        postId: postId,
-        body: body,
-      });
-      return res.json(newComment);
+      if (!validation) {
+        return res.status(404).json({ message: "Something went wrong" });
+      } else {
+        const newComment = await commentModel.create({
+          userId: userId,
+          postId: postId,
+          body: body,
+        });
+        return res.json(newComment);
+      }
     } catch (err) {
       console.log(err);
     }
@@ -26,7 +30,7 @@ export default {
       if (allComments.length !== 0) {
         return res.json(allComments);
       } else {
-        return res.status(404).send("No Comments Found");
+        return res.status(404).json({ message: "No Comments Found" });
       }
     } catch (err) {
       console.log(err);
@@ -35,15 +39,32 @@ export default {
 
   getUserComment: async (req, res) => {
     const { id, userId, postId } = req.body;
+    const validation = id && userId && postId;
     try {
-      const userComment = await commentModel.findOne({
-        where: { id: id, userId: userId, postId: postId },
-        include: [{ model: postModel }],
-      });
-      if (userComment) {
-        return res.json(userComment);
+      if (!validation) {
+        return res.status(404).json({ message: "Something went wrong" });
       } else {
-        return res.status(404).send("Comment Not Found");
+        const user = await userModel.findByPk(userId);
+        if (user) {
+          const post = await postModel.findByPk(postId);
+          if (post) {
+            const userComment = await commentModel.findOne({
+              where: { id: id, userId: userId, postId: postId },
+              include: [{ model: postModel }],
+            });
+            if (userComment) {
+              return res.json(userComment);
+            } else {
+              return res.status(404).json({ message: "Comment Not Found" });
+            }
+          } else {
+            return res
+              .status(404)
+              .json({ message: "Post Not Found Or Was Deleted" });
+          }
+        } else {
+          return res.status(404).json({ message: "User Not Found" });
+        }
       }
     } catch (err) {
       console.log(err);
@@ -53,16 +74,21 @@ export default {
   getAllUserComment: async (req, res) => {
     const { userId } = req.body;
     try {
-      const userComments = await commentModel.findAll({
-        where: { userId: userId },
-        include: [{ model: postModel }],
-      });
-      console.log(typeof userComments);
-      console.log(userComments);
-      if (userComments.length !== 0) {
-        return res.json(userComments);
+      const user = await userModel.findByPk(userId);
+      if (user) {
+        const userComments = await commentModel.findAll({
+          where: { userId: userId },
+          include: [{ model: postModel }],
+        });
+        if (userComments.length !== 0) {
+          return res.json(userComments);
+        } else {
+          return res
+            .status(404)
+            .json({ message: "The user dont have comments" });
+        }
       } else {
-        return res.status(404).send("The user dont have comments");
+        return res.status(404).json({ message: "User Not Found" });
       }
     } catch (err) {
       console.log(err);
@@ -72,15 +98,20 @@ export default {
   updateUserComment: async (req, res) => {
     const { id, userId, postId, body } = req.body;
     try {
-      const userComment = await commentModel.findOne({
-        where: { id: id, userId: userId, postId: postId },
-      });
-      if (userComment) {
-        userComment.body = body;
-        userComment.save();
-        return res.json(userComment);
+      const validation = id && userId && postId && body;
+      if (!validation) {
+        return res.status(404).json({ message: "Something went wrong" });
       } else {
-        return res.status(404).send("Comment Not Found");
+        const userComment = await commentModel.findOne({
+          where: { id: id, userId: userId, postId: postId },
+        });
+        if (userComment) {
+          userComment.body = body;
+          userComment.save();
+          return res.json(userComment);
+        } else {
+          return res.status(404).json({ message: "Comment Not Found" });
+        }
       }
     } catch (err) {
       console.log(err);
@@ -89,15 +120,22 @@ export default {
 
   deleteUserComment: async (req, res) => {
     const { id, userId, postId } = req.body;
+    const validation = id && userId && postId;
     try {
-      const userComment = await commentModel.findOne({
-        where: { id: id, userId: userId, postId: postId },
-      });
-      if (userComment) {
-        userComment.destroy();
-        return res.json("Comment successfully deleted");
+      if (!validation) {
+        return res.status(404).json({ message: "Something went wrong" });
       } else {
-        return res.status(404).send("Comment Not Found");
+        const userComment = await commentModel.findOne({
+          where: { id: id, userId: userId, postId: postId },
+        });
+        if (userComment) {
+          userComment.destroy();
+          return res
+            .status(200)
+            .json({ message: "Comment successfully deleted" });
+        } else {
+          return res.status(404).json({ message: "Comment Not Found" });
+        }
       }
     } catch (err) {
       console.log(err);
@@ -106,19 +144,26 @@ export default {
 
   deleteAllUserCommentsFromPost: async (req, res) => {
     const { userId, postId } = req.body;
+    const validation = userId && postId;
     try {
-      const userComments = await commentModel.findAll({
-        where: { userId: userId, postId: postId },
-      });
-      if (userComments.length !== 0) {
-        commentModel.destroy({
+      if (!validation) {
+        return res.status(404).json({ message: "Something went wrong" });
+      } else {
+        const userComments = await commentModel.findAll({
           where: { userId: userId, postId: postId },
         });
-        return res.json("All comments successfully deleted");
-      } else {
-        return res
-          .status(404)
-          .send("The user dont have comments on the specified post");
+        if (userComments.length !== 0) {
+          commentModel.destroy({
+            where: { userId: userId, postId: postId },
+          });
+          return res
+            .status(200)
+            .json({ message: "All comments successfully deleted" });
+        } else {
+          return res.status(404).json({
+            message: "The user dont have comments on the specified post",
+          });
+        }
       }
     } catch (err) {
       console.log(err);
@@ -128,16 +173,25 @@ export default {
   deleteAllUserComments: async (req, res) => {
     const { userId } = req.body;
     try {
-      const userComments = await commentModel.findAll({
-        where: { userId: userId },
-      });
-      if (userComments.length !== 0) {
-        commentModel.destroy({
+      const user = await userModel.findByPk(userId);
+      if (user) {
+        const userComments = await commentModel.findAll({
           where: { userId: userId },
         });
-        return res.json("All comments successfully deleted");
+        if (userComments.length !== 0) {
+          commentModel.destroy({
+            where: { userId: userId },
+          });
+          return res
+            .status(200)
+            .json({ message: "All comments successfully deleted" });
+        } else {
+          return res
+            .status(404)
+            .json({ message: "The user dont have comments" });
+        }
       } else {
-        return res.status(404).send("The user dont have comments");
+        return res.status(404).json({ message: "User Not Found" });
       }
     } catch (err) {
       console.log(err);
